Return Double instead of Float.
By returning the same value as a double we
avoid some awkward rounding problems when
using NumberFormatter for getting the va-
lue as a String.

+++ src/es/udc/cartolab/cit/gvsig/fmap/drivers/jdbc/spatialite/SpatiaLiteDriver.java
@@ -862,7 +862,7 @@ public class SpatiaLiteDriver extends DefaultJDBCDriver implements
 			if (metaData.getColumnType(fieldId) == Types.VARCHAR)
 				return ValueFactory.createValue(aRs.getString(fieldId));
 			if (metaData.getColumnType(fieldId) == Types.FLOAT)
-				return ValueFactory.createValue(aRs.getFloat(fieldId));
+				return ValueFactory.createValue(aRs.getDouble(fieldId));
 			if (metaData.getColumnType(fieldId) == Types.INTEGER)
 				return ValueFactory.createValue(aRs.getInt(fieldId));
 		}
